requirements.txt
numpy>=1.24
opencv-python>=4.9
Flask>=3.0
Pillow>=10.0
